Fix sinfo to print correct node state.
Bug 5505

+++ src/sinfo/print.c
@@ -1089,7 +1089,7 @@ int _print_state_compact(sinfo_data_t * sinfo_data, int width,
 			bool right_justify, char *suffix)
 {
 	char *upper_state, *lower_state;
-	uint16_t my_state;
+	uint32_t my_state;
 
 	if (sinfo_data && sinfo_data->nodes_total) {
 		my_state = sinfo_data->node_state;
@@ -1111,7 +1111,7 @@ int _print_state_long(sinfo_data_t * sinfo_data, int width,
 			bool right_justify, char *suffix)
 {
 	char *upper_state, *lower_state;
-	uint16_t my_state;
+	uint32_t my_state;
 
 	if (sinfo_data && sinfo_data->nodes_total) {
 		my_state = sinfo_data->node_state;
